added proper response for trasactions

+++ src/main/java/com/springBank/service/CustomerService.java
@@ -2,6 +2,7 @@ package com.springBank.service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -9,8 +10,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.springBank.exception.ResourceNotFoundException;
+import com.springBank.model.Account;
 import com.springBank.model.Customer;
 import com.springBank.repository.CustomerRepository;
+import com.springBank.repository.AccountRepository;
 
 @Service
 @Transactional
@@ -18,6 +21,9 @@ public class CustomerService {
 	
 	@Autowired
     private CustomerRepository customerRepository;
+	
+	@Autowired
+    private AccountRepository accountRepository;
 	
 	public void addCustomer(Customer customer) {
 		
@@ -43,7 +49,13 @@ public class CustomerService {
 	
 	public ResponseEntity<Customer> delete(long id)throws ResourceNotFoundException {
 		Customer customer = customerRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("account" +id));
-		customerRepository.delete(customer);
+		List<Account> accounts = accountRepository.findAll();
+        List<Account> customerAccounts = accounts.stream().filter(account1 -> account1.getCustomerId() == id).collect(Collectors.toList());
+        if (customerAccounts.size()>0) {
+            System.out.println("Accounts found for customerID " + id);
+            return null;
+        }
+        customerRepository.delete(customer);
 		return ResponseEntity.ok().build();
 	}
